cocos-api-sdk.js - v1.20.239: Added check on userId/userHash and sessionId/sessionHash, received on the payload of the accessToken. Added _hasPayloadData-function. Changed the login-function, expanding with locations.7 Improved handling of the _handleError-function, using new _getErrorsFromResponse-function. Updated the isAuthorized-function. Added methods here() and whoami(). Solved a issue which caused the callbackError to be executed twice when incorrect JSON was returned. dded methods getDeviceId(), getDeviceName(), getTopologyId() and getTopologyName(). Changed the _setCookie-function, in order to send a timestamp/age. Added method handleLocation. Added method setIAm(). Solved a issue with the loginExpired-callback.
cocos-application-sdk.js - v1.20.239:
Changed reload-handling.
Added function _handleLoginResponse.
Added handling for when a device or location is found/available, but not actived.
Added a defaultTextLib for the application to use as fallback.
The default languages for the application will come from getBrowserLanguage().
Added handling for the idDevice/deviceName and idTopology/topologyName during the login/authorisation-process.
Added method handleLocation() and addLocationOptions().
Added automatic handling for the iam-parameter.

cocos-javascript-functions.js - v1.20.223:
Added functions getDefaultLanguage() and getBrowserLanguage().

cocos-jquery-functions.js - v.1.20.204
Fixed the $.button-extension in order to be complatible with both FontAwesome 4 and FontAwesome 5.